requirements.txt
spacy>=3.7.0
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
beautifulsoup4>=4.12.0
requests>=2.31.0
streamlit>=1.32.0
praw>=7.7.0
geopy>=2.4.0
scipy>=1.12.0
python-dotenv>=1.0.0
